Refactor selection constraints
